Initial version of OOXML class
OOXML class should be the root of everything.  Excel class should extend it.

Initial OOXML class is already superior to our Excel class; We can check that it's a valid OOXML file now during loading, and will be able to easily check if it's a valid Excel file.
Also created some helper classes to deal with OOXML PartNames