Temp method to remove all ExternalReferences
To get to an MVP quickly, I hacked together a method that will remove *ALL* ExternalReferences.  I did this because 1. often times these broken Excel files have no legitimate external references and 2. it's an easy hack to get an MVP.

In terms of code quality, this is absolutely disgusting and should not be allowed to remain!  It should be pawned off into multiple different methods in different classes.

In addition, I had to add a getter to the OOXML class so we can discover the root tmp dir.  I believe we will need this anyways regardless of our little hack here.