Initial refactor of OOXML object
Refactoring OOXML to run in RAM rather than extracting to disk.  This is a fairly massive refactor.  In doing this, I came across the System.IO.Packaging.ZipPackage object, which actually does exactly what I was attempting to implement manually - this will make things FAR better and more flexible, as well as aid in fixing some bugs.  This will be a fairly large undertaking, but it will be significantly better on the other side.

In this first commit, I have massively refactored the OOXML class, added a couple helper classes, and done a couple other tweaks.